Reviewed-on: https://develop.vicara.co/vicara/vicara-api/pulls/771

+++ api/src/service/infrastructure/deployment.rs
@@ -1229,24 +1229,37 @@ pub async fn start_deployment(
 	)
 	.await?;
 
-	let kubeconfig = service::get_kubernetes_config_for_region(
-		connection,
-		&deployment.region,
-		config,
-	)
-	.await?;
+	let workspace = db::get_workspace_info(connection, workspace_id)
+		.await?
+		.status(500)?;
 
-	service::update_kubernetes_deployment(
-		workspace_id,
-		deployment,
-		&image_name,
-		digest.as_deref(),
-		deployment_running_details,
-		kubeconfig,
-		config,
-		request_id,
-	)
-	.await?;
+	if workspace.is_spam {
+		db::update_deployment_status(
+			connection,
+			deployment_id,
+			&DeploymentStatus::Running,
+		)
+		.await?;
+	} else {
+		let kubeconfig = service::get_kubernetes_config_for_region(
+			connection,
+			&deployment.region,
+			config,
+		)
+		.await?;
+
+		service::update_kubernetes_deployment(
+			workspace_id,
+			deployment,
+			&image_name,
+			digest.as_deref(),
+			deployment_running_details,
+			kubeconfig,
+			config,
+			request_id,
+		)
+		.await?;
+	}
 
 	Ok(())
 }
@@ -1290,30 +1303,44 @@ pub async fn update_deployment_image(
 	)
 	.await?;
 
-	let kubeconfig =
-		service::get_kubernetes_config_for_region(connection, region, config)
-			.await?;
+	let workspace = db::get_workspace_info(connection, workspace_id)
+		.await?
+		.status(500)?;
 
-	service::update_kubernetes_deployment(
-		workspace_id,
-		&Deployment {
-			id: deployment_id.clone(),
-			name: name.to_string(),
-			registry: registry.clone(),
-			image_tag: image_tag.to_string(),
-			status: DeploymentStatus::Pushed,
-			region: region.clone(),
-			machine_type: machine_type.clone(),
-			current_live_digest: Some(digest.to_string()),
-		},
-		image_name,
-		Some(digest),
-		deployment_running_details,
-		kubeconfig,
-		config,
-		request_id,
-	)
-	.await?;
+	if workspace.is_spam {
+		db::update_deployment_status(
+			connection,
+			deployment_id,
+			&DeploymentStatus::Running,
+		)
+		.await?;
+	} else {
+		let kubeconfig = service::get_kubernetes_config_for_region(
+			connection, region, config,
+		)
+		.await?;
+
+		service::update_kubernetes_deployment(
+			workspace_id,
+			&Deployment {
+				id: deployment_id.clone(),
+				name: name.to_string(),
+				registry: registry.clone(),
+				image_tag: image_tag.to_string(),
+				status: DeploymentStatus::Pushed,
+				region: region.clone(),
+				machine_type: machine_type.clone(),
+				current_live_digest: Some(digest.to_string()),
+			},
+			image_name,
+			Some(digest),
+			deployment_running_details,
+			kubeconfig,
+			config,
+			request_id,
+		)
+		.await?;
+	}
 
 	Ok(())
 }
@@ -1374,18 +1401,24 @@ pub async fn stop_deployment(
 	)
 	.await?;
 
-	let kubeconfig = service::get_kubernetes_config_for_region(
-		connection, region_id, config,
-	)
-	.await?;
+	let workspace = db::get_workspace_info(connection, workspace_id)
+		.await?
+		.status(500)?;
 
-	service::delete_kubernetes_deployment(
-		workspace_id,
-		deployment_id,
-		kubeconfig,
-		request_id,
-	)
-	.await?;
+	if !workspace.is_spam {
+		let kubeconfig = service::get_kubernetes_config_for_region(
+			connection, region_id, config,
+		)
+		.await?;
+
+		service::delete_kubernetes_deployment(
+			workspace_id,
+			deployment_id,
+			kubeconfig,
+			request_id,
+		)
+		.await?;
+	}
 
 	Ok(())
 }
@@ -1431,18 +1464,24 @@ pub async fn delete_deployment(
 	)
 	.await?;
 
-	let kubeconfig = service::get_kubernetes_config_for_region(
-		connection, region_id, config,
-	)
-	.await?;
+	let workspace = db::get_workspace_info(connection, workspace_id)
+		.await?
+		.status(500)?;
 
-	service::delete_kubernetes_deployment(
-		workspace_id,
-		deployment_id,
-		kubeconfig,
-		request_id,
-	)
-	.await?;
+	if !workspace.is_spam {
+		let kubeconfig = service::get_kubernetes_config_for_region(
+			connection, region_id, config,
+		)
+		.await?;
+
+		service::delete_kubernetes_deployment(
+			workspace_id,
+			deployment_id,
+			kubeconfig,
+			request_id,
+		)
+		.await?;
+	}
 
 	Ok(())
 }
